customer_portal: Removed own config, using core's
