+++ doc/customerportal.md
@@ -2,12 +2,12 @@
 
 ## Setup
 - Follow the [core setup](core.md#setup) to initiate the database
-- Copy `etc/customerportal.conf` to `customerportal.conf` and edit accordingly
+- Edit the `customer_portal` section in the config accordingly.
 - For web setup see [web.md](web.md#setup)
 
 ## Operation
 - Start API
 ```shell
-python -m stustapay.customer_portal -c customerportal.conf -vvv api
+python -m stustapay.customer_portal -c server.conf -vvv api
 ```
 - To start the web ui see [web.md](web.md#running)
