--- etc/customerportal.conf
@@ -1,8 +1,0 @@
-database:
-  host: null
-  user: null
-  password: null
-  dbname: "stustapay"
-
-customer_portal:
-    base_url: "localhost:8082"
