+++ etc/server.conf
@@ -15,3 +15,8 @@ administration:
 core:
     secret_key: "very secret"
     sumup_affiliate_key: "pretty secret"
+
+customer_portal:
+    base_url: "localhost:8082"
+    base_bon_url: "https://bon.stustapay.de/{bon_output_file}"
+    data_privacy_url: "https://stustapay.de/datenschutz"
