+++ stustapay/customer_portal/__main__.py
@@ -1,9 +1,9 @@
 import asyncio
 
 from stustapay.core.args import Parser
+from stustapay.core.config import read_config
 
 from . import server
-from .config import read_config
 
 
 def main():
